Revert "Dev rq 1654 1213"
