+++ apps/forms-flow-ai/forms-flow-web/src/apiManager/services/FOI/foiMasterDataServices.js
@@ -50,11 +50,7 @@ import {
   
   export const fetchFOIProgramAreaList = () => {    
     return (dispatch) => {
-      httpGETRequest(
-        API.FOI_GET_PROGRAMAREAS_FORUSER_API,
-        {},
-        UserService.getToken()
-      )
+      httpGETRequest(API.FOI_GET_PROGRAMAREAS_API, {}, UserService.getToken())
         .then((res) => {
           if (res.data) {
             const foiProgramAreaList = res.data;
