+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/IAO/AdvancedSearch/SearchComponent.js
@@ -42,9 +42,6 @@ const MenuProps = {
       width: 250,
     },
   },
-  MenuListProps: {
-    autoFocusItem: false,
-  },
 };
 
 const useStyles = makeStyles((theme) => ({
@@ -483,7 +480,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="unopened"
                           onChange={handleRequestStateChange}
                           checked={requestState.unopened.checked}
-                          color="success"
                         />
                       }
                       label="Unopened"
@@ -496,7 +492,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="callforrecords"
                           onChange={handleRequestStateChange}
                           checked={requestState.callforrecords.checked}
-                          color="success"
                         />
                       }
                       label="Call for Records"
@@ -509,7 +504,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="review"
                           onChange={handleRequestStateChange}
                           checked={requestState.review.checked}
-                          color="success"
                         />
                       }
                       label="Records Review"
@@ -522,7 +516,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="signoff"
                           onChange={handleRequestStateChange}
                           checked={requestState.signoff.checked}
-                          color="success"
                         />
                       }
                       label="Ministry Sign Off"
@@ -535,7 +528,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="closed"
                           onChange={handleRequestStateChange}
                           checked={requestState.closed.checked}
-                          color="success"
                         />
                       }
                       label="Closed"
@@ -565,7 +557,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="overdue"
                           onChange={handleRequestStatusChange}
                           checked={requestStatus.overdue}
-                          color="success"
                         />
                       }
                       label="Overdue"
@@ -578,7 +569,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="onTime"
                           onChange={handleRequestStatusChange}
                           checked={requestStatus.onTime}
-                          color="success"
                         />
                       }
                       label="On Time"
@@ -609,7 +599,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="personal"
                           onChange={handleRequestTypeChange}
                           checked={requestTypes.personal}
-                          color="success"
                         />
                       }
                       label="Personal"
@@ -622,7 +611,6 @@ const AdvancedSearch = ({ userDetail }) => {
                           name="general"
                           onChange={handleRequestTypeChange}
                           checked={requestTypes.general}
-                          color="success"
                         />
                       }
                       label="General"
@@ -754,7 +742,6 @@ const AdvancedSearch = ({ userDetail }) => {
                                 programArea.bcgovcode
                               ) > -1
                             }
-                            color="success"
                           />
                           <ListItemText
                             primary={programArea.name}
