+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/IAO/columns.js
@@ -32,7 +32,6 @@ const ProcessingTeamColumns = [
     headerAlign: "left",
     valueGetter: onBehalfFullName,
     sortable: false,
-    width: 180,
   },
   {
     field: "requestType",
