--- apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/Ministry/AdvancedSearch/SearchComponent.js
@@ -1,808 +1,0 @@
-import React, { useState, useContext } from "react";
-import "../../dashboard.scss";
-import { useSelector } from "react-redux";
-
-import Loading from "../../../../../containers/Loading";
-import Grid from "@mui/material/Grid";
-import Paper from "@mui/material/Paper";
-import InputBase from "@mui/material/InputBase";
-import IconButton from "@mui/material/IconButton";
-import Typography from "@mui/material/Typography";
-import TextField from "@mui/material/TextField";
-import Chip from "@mui/material/Chip";
-import SearchIcon from "@mui/icons-material/Search";
-import { makeStyles } from "@material-ui/core/styles";
-import FormGroup from "@mui/material/FormGroup";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import Checkbox from "@mui/material/Checkbox";
-import Button from "@mui/material/Button";
-import Menu from "@mui/material/Menu";
-import MenuItem from "@mui/material/MenuItem";
-import InputAdornment from "@mui/material/InputAdornment";
-
-import OutlinedInput from "@mui/material/OutlinedInput";
-import InputLabel from "@mui/material/InputLabel";
-import FormControl from "@mui/material/FormControl";
-import ListItemText from "@mui/material/ListItemText";
-import Select from "@mui/material/Select";
-import { SearchFilter } from "./enum";
-import {
-  ConditionalComponent,
-  formatDate,
-} from "../../../../../helper/FOI/helper";
-import { ActionContext } from "./ActionContext";
-import { StateEnum } from "../../../../../constants/FOI/statusEnum";
-
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-  MenuListProps: {
-    autoFocusItem: false,
-  },
-};
-
-const useStyles = makeStyles((theme) => ({
-  search: {
-    borderBottom: "1px solid #38598A",
-    backgroundColor: "rgba(56,89,138,0.1)",
-  },
-  searchBody: {
-    padding: "2em",
-  },
-  label: {
-    marginBottom: "1em",
-  },
-  chip: {
-    color: "#38598A",
-    border: "1px solid #38598A",
-    width: "100%",
-  },
-  paper: {
-    backgroundColor: "#F9FBFD",
-  },
-  checkboxLabel: {
-    marginBottom: 0,
-  },
-}));
-
-const AdvancedSearch = ({ userDetail }) => {
-  const classes = useStyles();
-
-  const {
-    handleUpdateSearchFilter,
-    searchLoading,
-    defaultSortModel,
-    advancedSearchComponentLoading,
-    setAdvancedSearchComponentLoading,
-    setSearchLoading,
-  } = useContext(ActionContext);
-
-  const programAreaList = useSelector(
-    (state) => state.foiRequests.foiProgramAreaList
-  );
-
-  const isLoading = useSelector((state) => state.foiRequests.isLoading);
-
-  const [searchText, setSearchText] = useState("");
-  const [keywords, setKeywords] = useState([]);
-  const [searchFilterSelected, setSearchFilterSelected] = useState(
-    SearchFilter.REQUEST_DESCRIPTION
-  );
-  const keywordsMode =
-    searchFilterSelected === SearchFilter.REQUEST_DESCRIPTION;
-
-  const intitialRequestState = {
-    callforrecords: {
-      checked: false,
-      id: StateEnum.callforrecords.id,
-    },
-    review: {
-      checked: false,
-      id: StateEnum.review.id,
-    },
-    signoff: {
-      checked: false,
-      id: StateEnum.signoff.id,
-    },
-    closed: {
-      checked: false,
-      id: StateEnum.closed.id,
-    },
-    callforrecordsoverdue: {
-      checked: false,
-      id: StateEnum.callforrecordsoverdue.id,
-    },
-  };
-  const [requestState, setRequestState] = useState(intitialRequestState);
-
-  const intitialRequestStatus = {
-    allActive: false,
-    overdue: false,
-    onTime: false,
-  };
-  const [requestStatus, setRequestStatus] = useState(intitialRequestStatus);
-
-  const initialRequestTypes = {
-    personal: false,
-    general: false,
-  };
-  const [requestTypes, setRequestTypes] = useState(initialRequestTypes);
-
-  const [fromDate, setFromDate] = useState("");
-  const [toDate, setToDate] = useState("");
-
-  const [selectedPublicBodies, setSelectedPublicBodies] = useState([]);
-
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const open = Boolean(anchorEl) && Boolean(searchText);
-
-  const getTrueKeysFromCheckboxObject = (checkboxObject) => {
-    return Object.entries(checkboxObject)
-      .map(([key, value]) => {
-        if (value instanceof Object) {
-          return value.checked ? value.id : null;
-        }
-        return value ? key.toLowerCase() : null;
-      })
-      .filter((value) => value);
-  };
-  const handleApplySearchFilters = () => {
-    if (!advancedSearchComponentLoading) {
-      setAdvancedSearchComponentLoading(true);
-    }
-    setSearchLoading(true);
-    handleUpdateSearchFilter({
-      search: searchFilterSelected.replace("_", "").toLowerCase(),
-      keywords: keywordsMode ? keywords : [searchText],
-      requestState: getTrueKeysFromCheckboxObject(requestState),
-      requestType: getTrueKeysFromCheckboxObject(requestTypes),
-      requestStatus: getTrueKeysFromCheckboxObject(requestStatus),
-      fromDate: fromDate || null,
-      toDate: toDate || null,
-      publicBodies: selectedPublicBodies,
-      page: 1,
-      size: 10,
-      sort: defaultSortModel,
-      userId: userDetail.preferred_username,
-    });
-  };
-
-  const handleResetSearchFilters = () => {
-    setSearchText("");
-    setKeywords([]);
-    setSearchFilterSelected(SearchFilter.REQUEST_DESCRIPTION);
-    setRequestState(intitialRequestState);
-    setRequestTypes(initialRequestTypes);
-    setRequestStatus(intitialRequestStatus);
-    setFromDate("");
-    setToDate("");
-    setSelectedPublicBodies([]);
-  };
-
-  const handleKeywordAdd = () => {
-    if (!searchText) {
-      return;
-    }
-    setAnchorEl(null);
-    setKeywords([...keywords, searchText]);
-    setSearchText("");
-  };
-
-  const handleSearchChange = (e) => {
-    setAnchorEl(e.currentTarget);
-    setSearchText(e.target.value);
-  };
-
-  const clickSearchFilter = (SearchFilterType) => {
-    if (searchFilterSelected !== SearchFilterType) {
-      setSearchFilterSelected(SearchFilterType);
-    }
-  };
-
-  const handleRequestStateChange = (event) => {
-    setRequestState({
-      ...requestState,
-      [event.target.name]: {
-        ...requestState[event.target.name],
-        checked: event.target.checked,
-      },
-    });
-  };
-
-  const handleRequestStatusChange = (event) => {
-    setRequestStatus({
-      ...requestStatus,
-      [event.target.name]: event.target.checked,
-    });
-  };
-
-  const handleRequestTypeChange = (event) => {
-    setRequestTypes({
-      ...requestTypes,
-      [event.target.name]: event.target.checked,
-    });
-  };
-
-  const handleSelectedPublicBodiesChange = (event) => {
-    const {
-      target: { value },
-    } = event;
-    setSelectedPublicBodies(
-      // On autofill we get a stringified value.
-      typeof value === "string" ? value.split(",") : value
-    );
-  };
-
-  const ClickableChip = ({ clicked, ...rest }) => {
-    if (!clicked) {
-      return (
-        <Chip
-          sx={[
-            {
-              color: "#38598A",
-              border: "1px solid #38598A",
-              width: "100%",
-            },
-          ]}
-          variant="outlined"
-          {...rest}
-        />
-      );
-    }
-
-    return (
-      <Chip
-        sx={[
-          {
-            backgroundColor: "#38598A",
-            width: "100%",
-          },
-          {
-            "&:hover": {
-              backgroundColor: "#38598A",
-            },
-          },
-        ]}
-        {...rest}
-      />
-    );
-  };
-
-  if (isLoading) {
-    return (
-      <Grid item xs={12} container alignItems="center">
-        <Loading costumStyle={{ position: "relative", marginTop: "4em" }} />
-      </Grid>
-    );
-  }
-  return (
-    <>
-      <Grid item container xs={12}>
-        <Grid item xs={12}>
-          <Paper
-            component={Grid}
-            sx={{
-              display: "flex",
-              width: "100%",
-              backgroundColor: "#F9FBFD",
-              border: "1px solid #38598A",
-            }}
-            alignItems="flex-start"
-            direction="row"
-            container
-          >
-            <Grid
-              item
-              container
-              alignItems="center"
-              direction="row"
-              xs={12}
-              className={classes.search}
-            >
-              <Grid item xs={keywordsMode ? 6 : 12}>
-                <InputBase
-                  placeholder="Search"
-                  onChange={handleSearchChange}
-                  value={searchText}
-                  onKeyPress={(e) => {
-                    if (keywordsMode && e.key === "Enter") {
-                      handleKeywordAdd();
-                    }
-                  }}
-                  startAdornment={
-                    <InputAdornment position="start">
-                      <IconButton sx={{ color: "#38598A" }}>
-                        <SearchIcon />
-                      </IconButton>
-                    </InputAdornment>
-                  }
-                  sx={{
-                    color: "#38598A",
-                  }}
-                  fullWidth
-                />
-              </Grid>
-              <ConditionalComponent condition={keywordsMode}>
-                <Grid item container direction="row-reverse" xs={6}>
-                  {keywords.map((keyword, index) => (
-                    <Chip
-                      key={`keyword-${index}`}
-                      label={keyword}
-                      onDelete={() => {
-                        setKeywords(keywords.filter((kw, i) => index !== i));
-                      }}
-                      color="primary"
-                      sx={{
-                        backgroundColor: "#38598A",
-                        margin: "1px",
-                      }}
-                      size="small"
-                    />
-                  ))}
-                </Grid>
-              </ConditionalComponent>
-            </Grid>
-            <ConditionalComponent condition={keywordsMode}>
-              <Grid>
-                <Menu
-                  id="basic-menu"
-                  anchorEl={anchorEl}
-                  open={open}
-                  onClose={handleKeywordAdd}
-                  disableAutoFocus={true}
-                  autoFocus={false}
-                >
-                  <MenuItem
-                    onClick={handleKeywordAdd}
-                  >{`Add "${searchText}"`}</MenuItem>
-                </Menu>
-              </Grid>
-            </ConditionalComponent>
-
-            <Grid
-              item
-              container
-              alignItems="flex-start"
-              justifyContent="flex-start"
-              direction="row"
-              xs={12}
-              className={classes.searchBody}
-              rowSpacing={2}
-              columnSpacing={1}
-            >
-              <Grid item xs={12}>
-                <Typography
-                  sx={{
-                    fontWeight: "bold",
-                  }}
-                  variant="h6"
-                >
-                  Advanced Search
-                </Typography>
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-request-description`}
-                  label={"REQUEST DESCRIPTION"}
-                  color="primary"
-                  onClick={() =>
-                    clickSearchFilter(SearchFilter.REQUEST_DESCRIPTION)
-                  }
-                  clicked={
-                    searchFilterSelected === SearchFilter.REQUEST_DESCRIPTION
-                  }
-                />
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-raw-request`}
-                  label={"ID NUMBER"}
-                  color="primary"
-                  onClick={() => clickSearchFilter(SearchFilter.ID_NUM)}
-                  clicked={searchFilterSelected === SearchFilter.ID_NUM}
-                />
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-axis-request`}
-                  label={"AXIS REQUEST #"}
-                  color="primary"
-                  onClick={() =>
-                    clickSearchFilter(SearchFilter.AXIS_REQUEST_NUM)
-                  }
-                  clicked={
-                    searchFilterSelected === SearchFilter.AXIS_REQUEST_NUM
-                  }
-                />
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-applicant-name`}
-                  label={"APPLICANT NAME"}
-                  color="primary"
-                  onClick={() => clickSearchFilter(SearchFilter.APPLICANT_NAME)}
-                  clicked={searchFilterSelected === SearchFilter.APPLICANT_NAME}
-                />
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-assignee-name`}
-                  label={"ASSIGNEE NAME"}
-                  color="primary"
-                  onClick={() => clickSearchFilter(SearchFilter.ASSIGNEE_NAME)}
-                  clicked={searchFilterSelected === SearchFilter.ASSIGNEE_NAME}
-                />
-              </Grid>
-
-              <Grid item xs={2}>
-                <ClickableChip
-                  key={`filter-search-filter`}
-                  label={"SUBJECT CODE"}
-                  color="primary"
-                  onClick={() => clickSearchFilter(SearchFilter.SUBJECT_CODE)}
-                  clicked={searchFilterSelected === SearchFilter.SUBJECT_CODE}
-                />
-              </Grid>
-
-              <Grid item xs={2} container direction="row" rowSpacing={2}>
-                <Grid item xs={12}>
-                  <Typography
-                    sx={{
-                      fontWeight: "bold",
-                    }}
-                    variant="h6"
-                  >
-                    Request State
-                  </Typography>
-                </Grid>
-                <Grid item xs={12}>
-                  <FormGroup>
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="callforrecords"
-                          onChange={handleRequestStateChange}
-                          checked={requestState.callforrecords.checked}
-                          color="success"
-                        />
-                      }
-                      label="Call for Records"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="review"
-                          onChange={handleRequestStateChange}
-                          checked={requestState.review.checked}
-                          color="success"
-                        />
-                      }
-                      label="Records Review"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="signoff"
-                          onChange={handleRequestStateChange}
-                          checked={requestState.signoff.checked}
-                          color="success"
-                        />
-                      }
-                      label="Ministry Sign Off"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="closed"
-                          onChange={handleRequestStateChange}
-                          checked={requestState.closed.checked}
-                          color="success"
-                        />
-                      }
-                      label="Closed"
-                    />
-                  </FormGroup>
-                </Grid>
-              </Grid>
-
-              <Grid item xs={2} container direction="row" rowSpacing={2}>
-                <Grid item xs={12}>
-                  <Typography
-                    sx={{
-                      fontWeight: "bold",
-                    }}
-                    variant="h6"
-                  >
-                    Request Status
-                  </Typography>
-                </Grid>
-                <Grid item xs={12}>
-                  <FormGroup>
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="allActive"
-                          onChange={handleRequestStatusChange}
-                          checked={requestStatus.allActive}
-                          color="success"
-                        />
-                      }
-                      label="All Active"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="overdue"
-                          onChange={handleRequestStatusChange}
-                          checked={requestStatus.overdue}
-                          color="success"
-                        />
-                      }
-                      label="Overdue"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="onTime"
-                          onChange={handleRequestStatusChange}
-                          checked={requestStatus.onTime}
-                          color="success"
-                        />
-                      }
-                      label="On Time"
-                    />
-                  </FormGroup>
-                </Grid>
-              </Grid>
-
-              <Grid item xs={2} container direction="row" rowSpacing={2}>
-                <Grid item xs={12}>
-                  <Typography
-                    sx={{
-                      fontWeight: "bold",
-                    }}
-                    variant="h6"
-                  >
-                    Request Type
-                  </Typography>
-                </Grid>
-
-                <Grid item xs={12}>
-                  <FormGroup>
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="personal"
-                          onChange={handleRequestTypeChange}
-                          checked={requestTypes.personal}
-                          color="success"
-                        />
-                      }
-                      label="Personal"
-                    />
-                    <FormControlLabel
-                      className={classes.checkboxLabel}
-                      control={
-                        <Checkbox
-                          size="small"
-                          name="general"
-                          onChange={handleRequestTypeChange}
-                          checked={requestTypes.general}
-                          color="success"
-                        />
-                      }
-                      label="General"
-                    />
-                  </FormGroup>
-                </Grid>
-              </Grid>
-
-              <Grid item xs={6} container direction="row" spacing={2}>
-                <Grid item xs={12}>
-                  <Typography
-                    sx={{
-                      fontWeight: "bold",
-                    }}
-                    variant="h6"
-                  >
-                    Search by Date Range
-                  </Typography>
-                </Grid>
-
-                <Grid
-                  container
-                  item
-                  direction="row"
-                  justifyContent="flex-start"
-                  alignItems="center"
-                  xs={12}
-                  sx={{
-                    mb: "1em",
-                  }}
-                >
-                  <Grid item xs={5}>
-                    <TextField
-                      id="from-date"
-                      label="From"
-                      type="date"
-                      InputLabelProps={{
-                        shrink: true,
-                      }}
-                      variant="outlined"
-                      value={fromDate || ""}
-                      InputProps={{
-                        inputProps: {
-                          max: formatDate(toDate) || formatDate(new Date()),
-                        },
-                      }}
-                      onChange={(e) => setFromDate(formatDate(e.target.value))}
-                      fullWidth
-                    />
-                  </Grid>
-                  <Grid item xs={2}>
-                    <Typography
-                      align="center"
-                      sx={{
-                        fontWeight: "bold",
-                      }}
-                      variant="h6"
-                    >
-                      to
-                    </Typography>
-                  </Grid>
-
-                  <Grid item xs={5}>
-                    <TextField
-                      id="to-date"
-                      label="To"
-                      type="date"
-                      InputLabelProps={{
-                        shrink: true,
-                      }}
-                      InputProps={{
-                        inputProps: {
-                          min: formatDate(fromDate),
-                          max: formatDate(new Date()),
-                        },
-                      }}
-                      value={toDate || ""}
-                      onChange={(e) => setToDate(formatDate(e.target.value))}
-                      variant="outlined"
-                      fullWidth
-                    />
-                  </Grid>
-                </Grid>
-
-                <Grid item xs={12}>
-                  <Typography
-                    sx={{
-                      fontWeight: "bold",
-                    }}
-                    variant="h6"
-                  >
-                    Search by Public Body
-                  </Typography>
-                </Grid>
-
-                <Grid item xs={12}>
-                  <FormControl fullWidth>
-                    <InputLabel id="multiple-checkbox-label" shrink>
-                      Public Body
-                    </InputLabel>
-                    <Select
-                      labelId="multiple-checkbox-label"
-                      id="multiple-checkbox"
-                      multiple
-                      displayEmpty
-                      value={selectedPublicBodies}
-                      onChange={handleSelectedPublicBodiesChange}
-                      input={<OutlinedInput label="Public Body" notched />}
-                      renderValue={(selected) => {
-                        if (programAreaList?.length === 1) {
-                          return programAreaList[0].bcgovcode;
-                        }
-                        if (selected.length === 0) {
-                          return <em>All</em>;
-                        }
-
-                        return selected.join(", ");
-                      }}
-                      MenuProps={MenuProps}
-                      disabled={programAreaList.length === 1}
-                    >
-                      <MenuItem disabled value="" key="program-area-all">
-                        <em>All</em>
-                      </MenuItem>
-                      {programAreaList.map((programArea) => (
-                        <MenuItem
-                          key={`program-area-${programArea.programareaid}`}
-                          value={programArea.bcgovcode}
-                        >
-                          <Checkbox
-                            checked={
-                              selectedPublicBodies.indexOf(
-                                programArea.bcgovcode
-                              ) > -1
-                            }
-                            color="success"
-                          />
-                          <ListItemText
-                            primary={programArea.name}
-                            key={`program-area-label-${programArea.programareaid}`}
-                          />
-                        </MenuItem>
-                      ))}
-                    </Select>
-                  </FormControl>
-                </Grid>
-              </Grid>
-
-              <Grid item xs={12} container direction="row" columnSpacing={1}>
-                <Grid item>
-                  <Button
-                    color="primary"
-                    sx={{
-                      backgroundColor: "#38598A",
-                      color: "white",
-                      fontWeight: "bold",
-                      textTransform: "none",
-                    }}
-                    variant="contained"
-                    onClick={handleApplySearchFilters}
-                    disabled={searchLoading}
-                    disableElevation
-                  >
-                    Apply Search
-                  </Button>
-                </Grid>
-                <Grid item>
-                  <Button
-                    variant="outlined"
-                    sx={{
-                      color: "#38598A",
-                      border: "1px solid #38598A",
-                      fontWeight: "bold",
-                      textTransform: "none",
-                    }}
-                    onClick={handleResetSearchFilters}
-                  >
-                    Clear All Filters
-                  </Button>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Paper>
-        </Grid>
-      </Grid>
-    </>
-  );
-};
-
-export default AdvancedSearch;
