--- apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/Ministry/AdvancedSearch/enum.js
@@ -1,9 +1,0 @@
-export const SearchFilter = Object.freeze({
-  REQUEST_DESCRIPTION: "REQUEST_DESCRIPTION",
-  RAW_REQUEST_NUM: "RAW_REQUEST_NUMBER",
-  ID_NUM: "ID_NUMBER",
-  AXIS_REQUEST_NUM: "AXIS_REQUEST_NUMBER",
-  APPLICANT_NAME: "APPLICANT_NAME",
-  ASSIGNEE_NAME: "ASSIGNEE_NAME",
-  SUBJECT_CODE: "SUBJECT_CODE",
-});
