--- apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/Ministry/AdvancedSearch/index.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { ActionProvider } from "./ActionContext";
-import SearchComponent from "./SearchComponent";
-import DataGridAdvancedSearch from "./DataGridAdvancedSearch";
-
-const AdvancedSearch = React.memo(({ userDetail }) => {
-  return (
-    <ActionProvider>
-      <SearchComponent userDetail={userDetail} />
-      <DataGridAdvancedSearch userDetail={userDetail} />
-    </ActionProvider>
-  );
-});
-
-export default AdvancedSearch;
