+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/index.js
@@ -1,4 +1,4 @@
 
 import Dashboard from "./IAO/Dashboard";
-import MinistryDashboard from "./Ministry/MinistryDashboard";
+import MinistryDashboard from "./MinistryDashboard";
 export { Dashboard, MinistryDashboard };
