+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/FOIRequest/MinistryAssignToDropdown.js
@@ -115,7 +115,7 @@ const MinistryAssignToDropdown  = React.memo(({requestState, requestDetails, min
                     variant="outlined"
                     fullWidth
                     required = {isMinistryCoordinator && requestState.toLowerCase() == StateEnum.callforrecords.name.toLowerCase() }
-                    disabled = {!isMinistryCoordinator || (requestState.toLowerCase() == StateEnum.closed.name.toLowerCase())}
+                    disabled = {!isMinistryCoordinator}
                     error={isMinistryCoordinator && selectedMinistryAssignedTo.toLowerCase().includes("unassigned")  && requestState.toLowerCase() == StateEnum.callforrecords.name.toLowerCase() }                    
                 >            
                     {getMenuItems()}
