+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/FOIRequest/MinistryReview/MinistryReview.js
@@ -40,7 +40,6 @@ import Loading from "../../../../containers/Loading";
 import ExtensionDetails from "./ExtensionDetails";
 import clsx from "clsx";
 import { getMinistryBottomTextMap } from "./utils";
-import DivisionalTracking from '../DivisionalTracking';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -431,43 +430,6 @@ const MinistryReview = React.memo(({ userDetail }) => {
 
   const requestNumber = requestDetails && requestDetails.idNumber;
 
-
-  const stateBox = (
-    requestState?.toLowerCase() == StateEnum.closed.name.toLowerCase() ?
-    (<span className="state-box">Closed</span>)
-    :
-    (
-      <StateDropDown
-        requestState={requestState}
-        updateStateDropDown={updateStateDropDown}
-        requestStatus={_requestStatus}
-        handleStateChange={handleStateChange}
-        isMinistryCoordinator={true}
-        isValidationError={isValidationError}
-      />
-    )
-  );
-
-  const divisions = requestDetails?.divisions?.length > 0 ? requestDetails.divisions : [];
-  const ministrycode = requestDetails?.selectedMinistries?.length > 0 ? requestDetails.selectedMinistries[0].code : '';
-  const divisionsBox = (
-    requestState?.toLowerCase() == StateEnum.closed.name.toLowerCase() ?
-    (
-      <DivisionalTracking
-        divisions={divisions}
-      />
-    )
-    :
-    (
-      <RequestTracking
-        pubmindivstagestomain={pubmindivstagestomain}
-        existingDivStages={divisions}
-        ministrycode={ministrycode}
-        createMinistrySaveRequestObject={createMinistrySaveRequestObject}
-      />
-    )
-  );
-
   return !isLoading &&
     requestDetails &&
     Object.keys(requestDetails).length !== 0 &&
@@ -481,7 +443,14 @@ const MinistryReview = React.memo(({ userDetail }) => {
             </h1>
           </div>
           <div className="foileftpaneldropdown">
-            {stateBox}
+            <StateDropDown
+              requestState={requestState}
+              updateStateDropDown={updateStateDropDown}
+              requestStatus={_requestStatus}
+              handleStateChange={handleStateChange}
+              isMinistryCoordinator={true}
+              isValidationError={isValidationError}
+            />
           </div>
 
           <div className="tab">
@@ -565,7 +534,14 @@ const MinistryReview = React.memo(({ userDetail }) => {
                           requestDetails={requestDetails}
                           requestState={requestState}
                         />
-                        {divisionsBox}
+                        <RequestTracking
+                          pubmindivstagestomain={pubmindivstagestomain}
+                          existingDivStages={requestDetails.divisions}
+                          ministrycode={
+                            requestDetails.selectedMinistries[0].code
+                          }
+                          createMinistrySaveRequestObject={createMinistrySaveRequestObject}
+                        />
                         {/* <RequestNotes /> */}
                         <BottomButtonGroup
                           requestState={requestState}
