+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/FOIRequest/MinistryReview/MinistryReview.scss
@@ -132,8 +132,3 @@
 .nodivstages {
     color:red;
 }
-
-.state-box {
-    padding-left:12px;
-    color:white;
-}
