+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/FOIRequest/MinistryReview/RequestHeader.js
@@ -8,7 +8,6 @@ import { useDispatch, useSelector } from "react-redux";
 import { fetchFOIFullAssignedToList } from "../../../../apiManager/services/FOI/foiMasterDataServices";
 import { Watcher } from '../../customComponents';
 import { useParams } from 'react-router-dom';
-import { StateEnum } from '../../../../constants/FOI/statusEnum';
 
 const RequestHeader = React.memo(({requestDetails, userDetail, handleMinistryAssignedToValue, createMinistrySaveRequestObject}) => {
 
@@ -49,15 +48,6 @@ const RequestHeader = React.memo(({requestDetails, userDetail, handleMinistryAss
     const headerText = _requestDetails.idNumber ? `Request #${_requestDetails.idNumber}` : FOI_COMPONENT_CONSTANTS.REVIEW_REQUEST;
     const assignedToValue = getFullName(assignedToList, _requestDetails);
 
-    const watcherBox = (
-        requestState?.toLowerCase() == StateEnum.closed.name.toLowerCase() ?
-        (<></>)
-        :
-        (
-            <Watcher watcherFullList={ministryAssignedToList} ministryId={ministryId} userDetail={userDetail} />
-        )
-      );
-
     return (
 
         <div className="foi-request-review-header-row1">
@@ -69,7 +59,7 @@ const RequestHeader = React.memo(({requestDetails, userDetail, handleMinistryAss
                 </div>
                 <div className="foi-request-review-header-col1-row" style={{marginTop:5+'px',display:'block'}}>
                   
-                    {watcherBox}
+                    <Watcher watcherFullList={ministryAssignedToList} ministryId={ministryId} userDetail={userDetail} />
                    
                 </div>
             </div>
