+++ request-management-api/request_api/services/programareaservice.py
@@ -6,8 +6,3 @@ class programareaservice:
         """ Returns the active records
         """
         return ProgramArea.getprogramareas()
-
-    def getprogramareasforministryuser(self, groups = None):
-        """ Returns the active records
-        """
-        return ProgramArea.getprogramareasforministryuser(groups)
