+++ request-management-api/request_api/utils/enums.py
@@ -96,8 +96,4 @@ class UserGroup(Enum):
     flex = "Flex Team"
     processing = "@processing"
     ministry = "@bcgovcode Ministry Team" 
-
-class RequestorType(Enum):
-    applicant = 1    
-    onbehalfof = 2
-    child = 3
+
